moved context, definition and types in a types folder
