+++ analyzer/src/local.rs
@@ -1,4 +1,4 @@
-use crate::types::Type;
+use crate::types::types::Type;
 
 /// A local variable.
 #[derive(Debug, Clone, PartialEq)]
@@ -6,7 +6,7 @@ pub struct Local {
     /// The name of the variable.
     pub name: String,
 
-    /// The type of the variable.
+    /// The types of the variable.
     pub ty: Type,
 
     /// Whether the variable has been initialized.
